+++ mobile-app/custom-styles/drawChartStyles.ts
@@ -1,0 +1,7 @@
+import {
+  StyleSheet
+} from 'react-native';
+
+export const drawChartStyles = StyleSheet.create({
+  container: { flex: 1, justifyContent: 'center', alignItems: 'center', backgroundColor: '#fff' },
+});
