+++ mobile-app/custom-styles/genderSwitchStyles.ts
@@ -1,0 +1,19 @@
+import {
+    StyleSheet
+} from 'react-native';
+
+export const genderSwitchStyles = StyleSheet.create({
+    label: {
+        fontSize: 11,
+        marginBottom: 5,
+        fontWeight: 'bold',
+    },
+    row: {
+        flexDirection: 'row',
+        alignItems: 'center',
+    },
+    option: {
+        fontSize: 13,
+        marginHorizontal: 10,
+    },
+});
