+++ mobile-app/custom-styles/nameInputStyles.ts
@@ -1,0 +1,20 @@
+import {
+    StyleSheet
+} from 'react-native';
+
+export const nameInputStyles = StyleSheet.create({
+    label: {
+        fontSize: 11,
+        marginBottom: 5,
+        fontWeight: 'bold',
+    },
+    input: {
+        height: 50,
+        borderColor: '#4A90E2',
+        borderWidth: 1,
+        borderRadius: 8,
+        paddingHorizontal: 15,
+        fontSize: 16,
+        backgroundColor: '#fff',
+    },
+});
