+++ mobile-app/custom-styles/newKundliPageStyes.ts
@@ -1,0 +1,13 @@
+import {
+    StyleSheet
+} from 'react-native';
+
+export const newKundliPageStyles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: '#fff',
+    },
+    content: {
+        padding: 20,
+    },
+});
