+++ mobile-app/custom-styles/savedKundalisPageStyles.ts
@@ -1,0 +1,17 @@
+import {
+    StyleSheet
+} from 'react-native';
+
+
+export const savedKundalisPageStyles = StyleSheet.create({
+    container: {
+        flex: 1,
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    text: {
+        fontSize: 24,
+        fontWeight: 'bold',
+        color: '#0000ff',
+    },
+});
