+++ mobile-app/package.json
@@ -24,7 +24,8 @@
     "react-native-safe-area-context": "^5.8.0",
     "react-native-svg": "^15.12.1",
     "react-native-tab-view": "^4.3.1",
-    "react-native-vector-icons": "^10.3.0"
+    "react-native-vector-icons": "^10.3.0",
+    "zustand": "^5.0.14"
   },
   "devDependencies": {
     "@react-native-community/cli": "latest",
