+++ mobile-app/pages/SavedKundalis.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import { View, Text, StyleSheet } from 'react-native';
+import { savedKundalisPageStyles as styles } from '../custom-styles/savedKundalisPageStyles';
 
 export default function SavedKundalis() {
   return (
@@ -9,15 +10,3 @@ export default function SavedKundalis() {
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  text: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    color: '#0000ff',
-  },
-});
