+++ mobile-app/utils/createKundliPage.ts
@@ -1,0 +1,3 @@
+export type CreateKundliProps = {
+  onNavigateToView?: () => void;
+};
