+++ mobile-app/utils/newKundliPage.ts
@@ -1,0 +1,3 @@
+export type NewKundliProps = {
+  onCreateChart?: () => void;
+};
